Make path and homedir related logic more robust
Add ability to pass OONI_HOME environment variable
